Refactor memory code.
